hi everyone these are the 3 fles that are uploaded here. these are the tasks performed by using functions.
in these are 3 programs we have solved 3 multiple problems. in first problem what i did, i did simplty print the course name and exicuted the function calling. in second program e made a functiojn that can calculate a cube of given number. if user will enter 3 as an input it will tell the cube of 3 = 27. and in last program we did swap the values by using function like x= 20 and y=27 and after swaping values it will show x=27 and y=20. these were three problems that is uploded here.